apps/core/pyproject.toml
[project]
name = "shipshare-core"
version = "0.1.0"
description = "Django core service for ShipShare identity, installations, and repository tracking."
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<6.0",
  "djangorestframework>=3.15.0",
  "cryptography>=44.0.0",
  "httpx>=0.28.0",
  "PyJWT>=2.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core_app", "github_integration", "identity"]
